pyproject.toml
[tool.poetry]
name = "flask-template"
version = "0.1.0"
description = "A project for flask framework"
authors = ["yuto_zh <545023318@qq.com>"]

[[tool.poetry.source]]
name = "ali"
url = "https://mirrors.aliyun.com/pypi/simple"

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
Flask = "1.1.2"
Flask-Migrate = "2.7.0"
Flask-Script = "^2.0.6"
Flask-SQLAlchemy = "^2.5.1"
Flask-Cors = "3.0.10"
Flask-Security = "3.0.0"
Flask-Cache = "0.13.1"
mkdocs = "^1.2.2"
python-dotenv = "^0.19.0"
dynaconf = "^3.1.7"
psycopg2 = "^2.9.1"
pylint = "^2.10.2"
flake8 = "^3.9.2"
pydantic-sqlalchemy = "^0.0.9"
gevent = "21.12.0"
marshmallow = "3.14.0"
pyjwt = "2.3.0"
email-validator = "1.1.3"
flask-restful = "0.3.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
